LayoutTests/platform/gtk/accessibility/name-from-label-expected.txt

Full Name: 
This tests that AtkObjects derive their names from associated labels.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS child.role is 'AXRole: entry'
PASS child.title is 'AXTitle: Full Name:'
PASS successfullyParsed is true

TEST COMPLETE
